Updated GameState with new Player functionality.
Currently, the player moves in one unchanging direction, and will loop
around the board.  Key presses do not change direction, will have to
investigate as to why that is.  Also, after adding a tail element, the
program crashes with an IndexOutOfBoundsException thrown.  It seems to be
triggered by updating, maybe in tailFollow()?